refactor(collavre_slack): move JavaScript and CSS to engine
- Move slack_integration.js to engines/collavre_slack/app/javascript/collavre_slack.js
- Add collavre_slack to GEM_ENGINES in build.cjs for gem support
- Create CSS file in collavre_slack engine for slack channel styles
- Update application.js to import collavre_slack engine
- Update layout to include collavre_slack stylesheet

This enables collavre_slack to be distributed as a standalone gem
without requiring host app modifications.
