+++ app/javascript/application.js
@@ -15,12 +15,12 @@ ActiveStorage.start()
 
 // Import Collavre engine modules (side-effect imports)
 import "collavre"
+import "collavre_slack"
 
 // Host app specific modules
 import "./firebase_config"
 import "./github_integration"
 import "./notion_integration"
-import "./slack_integration"
 import "./timezone_detection"
 import "./oauth_callback"
 import "./doorkeeper_token"
