+++ app/views/layouts/application.html.erb
@@ -36,6 +36,7 @@
     <%= stylesheet_link_tag "collavre/print", media: 'print' %>
     <%= stylesheet_link_tag "collavre/user_menu" %>
     <%= stylesheet_link_tag "collavre/popup" %>
+    <%= stylesheet_link_tag "collavre_slack/slack_integration" %>
 
     <%= javascript_include_tag "application", "data-turbo-track": "reload", defer: true, type: "module" %>
