+++ script/build.cjs
@@ -18,6 +18,7 @@ const fs = require('fs');
 // ============================================================================
 const GEM_ENGINES = [
     { name: 'collavre', mainEntry: 'collavre.js' },
+    { name: 'collavre_slack', mainEntry: 'collavre_slack.js' },
     // Add more engines here as needed:
     // { name: 'another_engine', mainEntry: 'another_engine.js' },
 ];
